Grafik Perkembangan

+++ application/views/setting/balita.php
@@ -6,14 +6,21 @@
                     <h3>Data Balita</h3>
                 </div>
                 <div class="widget-content">
-                    <div class="clearfix">
-                        <div class="control-group pull-right">											
-                            <a href="<?= base_url()?>setting/balita_edit" role="button" class="btn btn-primary" data-backdrop="static">
-                                    <span class="icon-plus"></span> Tambah Baru
-                                </a>				
-                        </div>
-                       </div>
-                    <br>
+                    <?php
+                    if ($this->session->userdata('role_id') != '2')
+                    {
+                    ?>
+                        <div class="clearfix">
+                            <div class="control-group pull-right">											
+                                <a href="<?= base_url()?>setting/balita_edit" role="button" class="btn btn-primary" data-backdrop="static">
+                                        <span class="icon-plus"></span> Tambah Baru
+                                    </a>				
+                            </div>
+                           </div>
+                        <br>
+                    <?php
+                    }
+                    ?>
                     <table id="auditor_main" class="table table-striped table-bordered">
                     <thead>
                       <tr>
@@ -39,8 +46,15 @@
                                 <td><?= $r['berat_lahir']?></td>
                                 <td><?= $r['tinggi_lahir']?></td>
                                 <td>
+                                    <?php
+                                    if ($this->session->userdata('role_id') != '2')
+                                    {
+                                    ?>
                                     <button class="btn btn-info btn-small" href="<?= base_url()?>setting/petugas_new/<?= $r['id_balita']?>" ><i class="icon-edit"> </i></button>
                                     <button class="btn btn-danger btn-small" href="<?= base_url()?>setting/petugas_delete/<?= $r['id_balita']?>" ><i class="icon-remove-sign"> </i></button>
+                                    <?php
+                                    }
+                                    ?>
                                 </td>
                             </tr>
                         <?php
